organize imports, clean dependencies and add SAML views to cas-server-webapp
