+++ cas-server-support-saml/src/test/java/org/jasig/cas/support/saml/authentication/DirectMappingAuthenticationManagerImplTests.java
@@ -24,6 +24,8 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
+import junit.framework.TestCase;
+
 import org.jasig.cas.authentication.Authentication;
 import org.jasig.cas.authentication.AuthenticationMetaDataPopulator;
 import org.jasig.cas.authentication.DirectMappingAuthenticationManagerImpl;
@@ -34,8 +36,6 @@ import org.jasig.cas.authentication.principal.Credentials;
 import org.jasig.cas.authentication.principal.HttpBasedServiceCredentials;
 import org.jasig.cas.authentication.principal.UsernamePasswordCredentials;
 import org.jasig.cas.authentication.principal.UsernamePasswordCredentialsToPrincipalResolver;
-
-import junit.framework.TestCase;
 
 
 public class DirectMappingAuthenticationManagerImplTests extends TestCase {
