+++ cas-server-support-saml/src/test/java/org/jasig/cas/support/saml/web/view/Saml10SuccessResponseViewTests.java
@@ -26,11 +26,12 @@ import java.util.List;
 import java.util.Map;
 
 import junit.framework.TestCase;
+
 import org.jasig.cas.TestUtils;
 import org.jasig.cas.authentication.Authentication;
 import org.jasig.cas.authentication.MutableAuthentication;
-import org.jasig.cas.support.saml.authentication.SamlAuthenticationMetaDataPopulator;
 import org.jasig.cas.authentication.principal.SimplePrincipal;
+import org.jasig.cas.support.saml.authentication.SamlAuthenticationMetaDataPopulator;
 import org.jasig.cas.validation.Assertion;
 import org.jasig.cas.validation.ImmutableAssertionImpl;
 import org.springframework.mock.web.MockHttpServletRequest;
